+++ src/main/java/eu/fraho/libs/beencode/BInteger.java
@@ -6,42 +6,80 @@
  */
 package eu.fraho.libs.beencode;
 
+import org.jetbrains.annotations.Contract;
+import org.jetbrains.annotations.NotNull;
+
 import java.io.IOException;
 import java.io.InputStream;
 import java.io.OutputStream;
 import java.math.BigInteger;
-import java.util.Objects;
 
 public final class BInteger extends BNodeBase<BigInteger> implements Comparable<BInteger> {
     private static final long serialVersionUID = 100L;
-    private static final int MAX_READ_LEN = 31;
+    // maximum length of 64 bit signed integer is 20 chars + suffix
+    private static final int MAX_READ_LEN = 21;
     private static final byte PREFIX = 'i';
     private static final byte SUFFIX = 'e';
 
-    private BInteger(BigInteger value) {
+    private BInteger(@NotNull BigInteger value) {
         super(value);
     }
 
-    public static BInteger of(Integer value) {
-        Objects.requireNonNull(value, "value may not be null");
+    /**
+     * Create a new instance of the given value
+     *
+     * @param value value
+     * @return new instance
+     */
+    @Contract(pure = true, value = "_ -> new")
+    public static @NotNull BInteger of(@NotNull Integer value) {
         return of(BigInteger.valueOf(value));
     }
 
-    public static BInteger of(Long value) {
-        Objects.requireNonNull(value, "value may not be null");
+    /**
+     * Create a new instance of the given value
+     *
+     * @param value value
+     * @return new instance
+     */
+    @Contract(pure = true, value = "_ -> new")
+    public static @NotNull BInteger of(@NotNull Long value) {
         return of(BigInteger.valueOf(value));
     }
 
-    public static BInteger of(BigInteger value) {
-        Objects.requireNonNull(value, "value may not be null");
+    /**
+     * Create a new instance of the given value
+     *
+     * @param value value
+     * @return new instance
+     */
+    @Contract(pure = true, value = "_ -> new")
+    public static @NotNull BInteger of(@NotNull BigInteger value) {
         return new BInteger(value);
     }
 
-    public static BInteger of(InputStream is) throws IOException {
+    /**
+     * Parse the given stream for a BInteger
+     *
+     * @param is stream of data
+     * @return new instance
+     * @throws BencodeException if the given prefix is not {@link #PREFIX} or the parsed stream is invalid
+     */
+    @Contract(pure = true, value = "_ -> new")
+    public static @NotNull BInteger of(@NotNull InputStream is) throws IOException {
         return of(is, (byte) is.read());
     }
 
-    public static BInteger of(InputStream is, byte prefix) throws IOException {
+    /**
+     * Parse the given stream for a BInteger
+     *
+     * @param is     stream of data
+     * @param prefix first read byte, has to be {@link #PREFIX}
+     * @return new instance
+     * @throws BencodeException if the given prefix is not {@link #PREFIX} or the parsed stream is invalid
+     */
+    @Contract(value = "_, _ -> new")
+    public static @NotNull BInteger of(@NotNull InputStream is, byte prefix) throws IOException {
         if (!canParsePrefix(prefix)) {
             throw new BencodeException("Unknown prefix, cannot parse: " + prefix);
         }
@@ -50,7 +88,9 @@ public final class BInteger extends BNodeBase<BigInteger> implements Comparable<
         byte read = 0;
         for (int i = 0; i < MAX_READ_LEN; i++) {
             read = (byte) is.read();
-            if (read == SUFFIX) break;
+            if (read == SUFFIX) {
+                break;
+            }
             str.append((char) read);
         }
 
@@ -66,11 +106,7 @@ public final class BInteger extends BNodeBase<BigInteger> implements Comparable<
         if (input.equals("-")) {
             throw new BencodeException("Invalid data, only a dash was read");
         }
-        if (input.equals("-0")) {
-            throw new BencodeException("Invalid data, negative zero is not allowed");
-        }
-        if (input.startsWith("0") && length > 1 || input.startsWith("-0")
-                                                   && length > 2) {
+        if (input.startsWith("0") && length > 1 || input.startsWith("-0")) {
             throw new BencodeException("Invalid data, leading zeros are not allowed");
         }
 
@@ -81,29 +117,38 @@ public final class BInteger extends BNodeBase<BigInteger> implements Comparable<
         }
     }
 
+    /**
+     * @param prefix the prefix to check
+     * @return is the given byte the expected prefix for this element?
+     * @see #PREFIX
+     */
+    @Contract(pure = true)
     public static boolean canParsePrefix(byte prefix) {
         return prefix == PREFIX;
     }
 
     @Override
-    public void write(OutputStream os) throws IOException {
+    public void write(@NotNull OutputStream os) throws IOException {
         os.write(PREFIX);
         os.write(getValue().toString().getBytes(DEFAULT_CHARSET));
         os.write(SUFFIX);
     }
 
     @Override
-    public int compareTo(BInteger o) {
-        Objects.requireNonNull(o, "other is null");
+    @Contract(pure = true)
+    public int compareTo(@NotNull BInteger o) {
         return getValue().compareTo(o.getValue());
     }
 
+    /**
+     * Clone this object
+     *
+     * @return copy of this object
+     */
+    @SuppressWarnings("MethodDoesntCallSuperMethod")
     @Override
-    public BInteger clone() {
-        try {
-            return (BInteger) super.clone();
-        } catch (BencodeException be) {
-            return BInteger.of(getValue());
-        }
+    @Contract(pure = true, value = "-> new")
+    public @NotNull BInteger clone() {
+        return BInteger.of(getValue());
     }
 }