+++ src/main/java/eu/fraho/libs/beencode/BNode.java
@@ -6,6 +6,10 @@
  */
 package eu.fraho.libs.beencode;
 
+import org.jetbrains.annotations.Contract;
+import org.jetbrains.annotations.NotNull;
+import org.jetbrains.annotations.VisibleForTesting;
+
 import java.io.IOException;
 import java.io.OutputStream;
 import java.io.Serializable;
@@ -15,9 +19,16 @@ import java.nio.charset.StandardCharsets;
 public interface BNode<T> extends Cloneable, Serializable {
     Charset DEFAULT_CHARSET = StandardCharsets.US_ASCII;
 
-    void write(OutputStream os) throws IOException;
+    /**
+     * Write this element to the given stream.
+     *
+     * @param os the stream to write to
+     * @throws IOException if the write operation fails
+     */
+    void write(@NotNull OutputStream os) throws IOException;
 
+    @VisibleForTesting
+    @NotNull
+    @Contract(pure = true)
     T getValue();
-
-    BNode<T> clone();
 }