+++ src/main/java/eu/fraho/libs/beencode/NodeFactory.java
@@ -6,6 +6,9 @@
  */
 package eu.fraho.libs.beencode;
 
+import org.jetbrains.annotations.Contract;
+import org.jetbrains.annotations.NotNull;
+
 import java.io.*;
 import java.util.Objects;
 import java.util.Optional;
@@ -15,8 +18,8 @@ public final class NodeFactory {
         // this util class should not be instantiated
     }
 
-    public static BNode<?> decode(InputStream stream, byte prefix) throws IOException {
-        Objects.requireNonNull(stream, "stream may not be null");
+    @Contract(value = "_, _ -> new")
+    public static @NotNull BNode<?> decode(@NotNull InputStream stream, byte prefix) throws IOException {
         if (BDict.canParsePrefix(prefix)) {
             return BDict.of(stream, prefix);
         } else if (BInteger.canParsePrefix(prefix)) {
@@ -30,14 +33,13 @@ public final class NodeFactory {
         }
     }
 
-    public static BNode<?> decode(InputStream stream) throws IOException {
+    @Contract(value = "_ -> new")
+    public static @NotNull BNode<?> decode(@NotNull InputStream stream) throws IOException {
         Objects.requireNonNull(stream, "stream may not be null");
         return decode(stream, (byte) stream.read());
     }
 
-    public static <T extends BNode<?>> Optional<T> decode(InputStream stream, Class<T> expected) throws IOException {
-        Objects.requireNonNull(stream, "stream may not be null");
-        Objects.requireNonNull(expected, "expected may not be null");
+    public static <T extends BNode<?>> Optional<T> decode(@NotNull InputStream stream, @NotNull Class<T> expected) throws IOException {
         BNode<?> result = decode(stream, (byte) stream.read());
         if (expected.isAssignableFrom(result.getClass())) {
             return Optional.of(expected.cast(result));
@@ -46,7 +48,8 @@ public final class NodeFactory {
         }
     }
 
-    public static BNode<?> decode(byte[] data) {
+    @Contract(pure = true, value = "_ -> new")
+    public static @NotNull BNode<?> decode(byte @NotNull [] data) {
         Objects.requireNonNull(data, "data may not be null");
         try (InputStream is = new ByteArrayInputStream(data)) {
             return decode(is);
@@ -56,9 +59,8 @@ public final class NodeFactory {
         }
     }
 
-    public static <T extends BNode<?>> Optional<T> decode(byte[] data, Class<T> expected) {
-        Objects.requireNonNull(data, "data may not be null");
-        Objects.requireNonNull(expected, "expected may not be null");
+    @Contract(pure = true, value = "_, _ -> new")
+    public static <T extends BNode<?>> Optional<T> decode(byte @NotNull [] data, @NotNull Class<T> expected) {
         try (InputStream stream = new ByteArrayInputStream(data)) {
             return decode(stream, expected);
         } catch (IOException e) {
@@ -67,8 +69,8 @@ public final class NodeFactory {
         }
     }
 
-    public static byte[] encode(BNode<?> node) {
-        Objects.requireNonNull(node, "node may not be null");
+    @Contract(pure = true, value = "_ -> new")
+    public static byte @NotNull [] encode(@NotNull BNode<?> node) {
         try (ByteArrayOutputStream bos = new ByteArrayOutputStream()) {
             node.write(bos);
             return bos.toByteArray();
@@ -78,9 +80,7 @@ public final class NodeFactory {
         }
     }
 
-    public static void encode(BNode<?> node, OutputStream os) throws IOException {
-        Objects.requireNonNull(node, "node may not be null");
-        Objects.requireNonNull(os, "os may not be null");
+    public static void encode(@NotNull BNode<?> node, @NotNull OutputStream os) throws IOException {
         node.write(os);
     }
 }